[RF] Refactor MarketFactory to make variables immutable + save storage

+++ contracts/interfaces/IPMarketFactory.sol
@@ -21,7 +21,7 @@ interface IPMarketFactory {
         view
         returns (
             address treasury,
-            uint96 lnFeeRateRoot,
+            uint88 lnFeeRateRoot,
             uint8 reserveFeePercent
         );
 }
